samformer_pytorch/requirements.txt
pytorch==1.12.1
tqdm==4.64.1
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.1